+++ Lib/fontTools/misc/plistlib.py
@@ -541,44 +541,3 @@ def dumps(
         pretty_print=pretty_print,
     )
     return fp.getvalue()
-
-
-# The following functions were part of the old py2-like ufoLib.plistlib API.
-# They are kept only for backward compatiblity.
-from fontTools.ufoLib.utils import deprecated
-
-
-@deprecated("Use 'load' instead")
-def readPlist(path_or_file):
-    did_open = False
-    if isinstance(path_or_file, basestring):
-        path_or_file = open(path_or_file, "rb")
-        did_open = True
-    try:
-        return load(path_or_file, use_builtin_types=False)
-    finally:
-        if did_open:
-            path_or_file.close()
-
-
-@deprecated("Use 'dump' instead")
-def writePlist(value, path_or_file):
-    did_open = False
-    if isinstance(path_or_file, basestring):
-        path_or_file = open(path_or_file, "wb")
-        did_open = True
-    try:
-        dump(value, path_or_file, use_builtin_types=False)
-    finally:
-        if did_open:
-            path_or_file.close()
-
-
-@deprecated("Use 'loads' instead")
-def readPlistFromString(data):
-    return loads(tobytes(data, encoding="utf-8"), use_builtin_types=False)
-
-
-@deprecated("Use 'dumps' instead")
-def writePlistToString(value):
-    return dumps(value, use_builtin_types=False)
